api/video/requirements.txt
ffmpeg
moviepy
fastapi[all]==0.99.0
openai
av
boto3
python-dotenv
pydub